refactor: move Datum out of xy_chart
