+++ src/utils/domain.ts
@@ -105,3 +105,5 @@ export function computeStackedContinuousDomain(
   const cumulativeSumAccessor = (d: any) => d.value;
   return computeContinuousDataDomain(groups, cumulativeSumAccessor, scaleToExtent);
 }
+
+export type Datum = any;
